database items metrics moved to vmaas_sync
